requirements.txt
asgiref==3.3.4
defusedxml==0.7.1
diff-match-patch==20200713
django-import-export==2.5.0
django-simpleui==2021.5.21
Django==3.2.3
et-xmlfile==1.1.0
MarkupPy==1.14
mysqlclient==2.0.3
odfpy==1.4.1
openpyxl==3.0.7
pip==21.1.1
pytz==2021.1
PyYAML==5.4.1
setuptools==56.2.0
sqlparse==0.4.1
tablib==3.0.0
xlrd==2.0.1
xlwt==1.3.0
